+++ WordCount/WordCountAPI/Controllers/JsonInputModels/JsonSchemaInputModel.cs
