+++ WordCount/WordCountAPI/Data/DataAccess/IFilterBuilder.cs
