+++ WordCount/WordCountAPI/Data/Models/Publisher.cs
